Fix the sigaction declaration to match the system for Linux. I don't know whether Darwin needs a change here.
R=rsc
http://go/go-review/1013010

+++ src/pkg/runtime/linux/os.h
@@ -7,4 +7,4 @@ int32	futex(uint32*, int32, uint32, Timespec*, uint32*, uint32);
 int32	clone(int32, void*, M*, G*, void(*)(void));
 
 struct Sigaction;
-void	rt_sigaction(int64, struct Sigaction*, void*, uint64);
+void	rt_sigaction(uintptr, struct Sigaction*, void*, uintptr);
